requirements.txt
spotipy
syrics
pylrc
PyQt5
pyautogui
pillow